enable backend to share code within the monorepo
