+++ backend/package.json
@@ -19,13 +19,12 @@
   "main": "src/server.ts",
   "scripts": {
     "clean": "rm -f ../logs/adminActivity.log",
-    "start": "NODE_ENV=production node ./dist/server.js --log=1 --registry=localhost:50051",
-    "start:dev": "npm run clean && export NODE_TLS_REJECT_UNAUTHORIZED=0 && export NODE_ENV=development && nodemon src/server.ts --log=1 --registry=localhost:50051",
+    "start": "NODE_ENV=production node ./dist/server.js",
+    "start:dev": "npm run clean && NODE_TLS_REJECT_UNAUTHORIZED=0 NODE_ENV=development nodemon src/server.ts",
+    "debug": "NODE_TLS_REJECT_UNAUTHORIZED=0 NODE_ENV=development nodemon --exec 'node --inspect --require ts-node/register' src/server.ts",
     "start:dev:wait": "npx wait-on -i 1000 http://localhost:4000",
-    "debug": "npm run tsc && export NODE_TLS_REJECT_UNAUTHORIZED=0 && export NODE_ENV=development && node --inspect ./dist/server.js --log=1 --registry=localhost:50051",
-    "build-only": "tsc -p . && node ./dist/server.js --log=1 --registry=localhost:50051 --buildonly",
-    "build": "run-s build:clean tsc:prod",
-    "build:clean": "rimraf ./dist",
+    "build": "tsdown",
+    "build:watch": "tsdown --watch",
     "test": "run-s test:lint test:type-check test:jest",
     "test:unit-coverage": "npm run test:jest:coverage",
     "test:jest:coverage": "rimraf jest-coverage && jest --silent --coverage --coverageReporters=json --coverageReporters=lcov",
@@ -35,9 +34,6 @@
     "test:jest": "jest",
     "type-check": "tsc --noEmit",
     "server": "NODE_ENV=production node ./dist/server.js",
-    "tsc": "tsc -p .",
-    "tsc:prod": "tsc -p tsconfig.prod.json",
-    "watch": "tsc -p . -w",
     "lint": "eslint .",
     "test-unit": "jest --silent",
     "test-unit-coverage": "jest --silent --coverage"
@@ -51,6 +47,7 @@
     "@fastify/view": "^9.1.0",
     "@fastify/websocket": "^8.3.1",
     "@kubernetes/client-node": "^0.12.2",
+    "@odh-dashboard/app-config": "*",
     "dotenv": "^16.4.7",
     "ejs": "^3.1.10",
     "fastify": "^4.28.1",
@@ -61,20 +58,19 @@
     "lodash": "^4.17.21",
     "pino": "^9.9.0",
     "prom-client": "^14.0.1",
-    "ts-node": "^10.9.1"
+    "ws": "^8.17.1"
   },
   "devDependencies": {
     "@odh-dashboard/eslint-config": "*",
     "@types/http-errors": "^1.8.0",
     "@types/lodash": "^4.14.182",
     "@types/node": "^18.14.5",
-    "fastify-cli": "^5.7.1",
     "globals": "^16.4.0",
     "nodemon": "^3.0.2",
     "pino-pretty": "^9.4.0",
     "rimraf": "^3.0.2",
-    "ts-protoc-gen": "^0.10.0",
-    "tslib": "^2.0.3",
+    "ts-node": "^10.9.2",
+    "tsdown": "^0.15.6",
     "typescript": "^5.8.3"
   },
   "optionalDependencies": {
