+++ backend/src/routes/root.ts
@@ -1,12 +1,19 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
-import { getModuleFederationConfig } from './module-federation';
+import { getModuleFederationConfigs } from '@odh-dashboard/app-config';
+import { DEV_MODE } from '../utils/constants';
+import { errorHandler } from '../utils';
 
 export default async (fastify: FastifyInstance): Promise<void> => {
-  const remotes = getModuleFederationConfig(fastify)?.map((c) => ({
-    name: c.name,
-    remoteEntry: c.remoteEntry,
-  }));
-  const mfRemotesJson = remotes ? JSON.stringify(remotes) : undefined;
+  let mfRemotesJson: string;
+  try {
+    const remotes = getModuleFederationConfigs(DEV_MODE).map((c) => ({
+      name: c.name,
+      remoteEntry: c.remoteEntry,
+    }));
+    mfRemotesJson = remotes ? JSON.stringify(remotes) : undefined;
+  } catch (e) {
+    fastify.log.error(e, errorHandler(e));
+  }
 
   fastify.get('/*', async (_: FastifyRequest, reply: FastifyReply) =>
     reply.view('index.html', { mfRemotesJson }),
