+++ backend/src/utils/resourceUtils.ts
@@ -6,6 +6,7 @@ import {
   BuildKind,
   BuildStatus,
   ConsoleLinkKind,
+  CSVKind,
   DashboardConfig,
   K8sResourceCommon,
   KubeFastifyInstance,
@@ -30,7 +31,8 @@ import {
 } from './resourceWatcher';
 import { getComponentFeatureFlags } from './features';
 import { blankDashboardCR } from './constants';
-import { getIsAppEnabled, getRouteForApplication, getRouteForClusterId } from './componentUtils';
+import { getLink, getRouteForClusterId, getServiceLink } from './componentUtils';
+import { isHttpError } from '../utils';
 import { getDetectedAccelerators } from '../routes/api/accelerators/acceleratorUtils';
 import { FastifyRequest } from 'fastify';
 import { fetchClusterStatus } from './dsc';
@@ -617,6 +619,180 @@ export const getBuildStatuses = (): BuildStatus[] => {
 
 export const getConsoleLinks = (): ConsoleLinkKind[] => {
   return consoleLinksWatcher.getResources();
+};
+
+const getConsoleLinkRoute = (appDef: OdhApplication): string => {
+  if (!appDef.spec.consoleLink) {
+    return null;
+  }
+  const consoleLink = getConsoleLinks().find((cl) => cl.metadata.name === appDef.spec.consoleLink);
+  return consoleLink ? consoleLink.spec.href : null;
+};
+
+export const checkJupyterEnabled = (): boolean =>
+  getDashboardConfig().spec.notebookController?.enabled !== false;
+
+export const getApplicationEnabledConfigMap = async (
+  fastify: KubeFastifyInstance,
+  appDef: OdhApplication,
+): Promise<boolean> => {
+  const namespace = fastify.kube.namespace;
+  const name = appDef.spec.enable?.validationConfigMap;
+  if (!name) {
+    return Promise.resolve(null);
+  }
+  const coreV1Api = fastify.kube.coreV1Api;
+  const enabledCM = await coreV1Api
+    .readNamespacedConfigMap(name, namespace)
+    .then((result) => result.body)
+    .catch(() => null);
+  if (!enabledCM) {
+    return false;
+  }
+  return enabledCM.data?.validation_result === 'true';
+};
+
+const getField = (obj: any, path: string, defaultValue: string = undefined): string => {
+  const travel = (regexp: RegExp) =>
+    String.prototype.split
+      .call(path, regexp)
+      .filter(Boolean)
+      .reduce((res: any, key: string) => (res !== null && res !== undefined ? res[key] : res), obj);
+  const result = travel(/[,[\]]+?/) || travel(/[,[\].]+?/);
+  return result === undefined || result === obj ? defaultValue : result;
+};
+
+const getCREnabledForApp = (
+  fastify: KubeFastifyInstance,
+  appDef: OdhApplication,
+): Promise<boolean> => {
+  const { enableCR } = appDef.spec;
+  if (!enableCR) {
+    return undefined;
+  }
+
+  const customObjectsApi = fastify.kube.customObjectsApi;
+  const namespace = enableCR.namespace || fastify.kube.namespace;
+  const { group, version, plural, name } = enableCR;
+  return customObjectsApi
+    .getNamespacedCustomObject(group, version, namespace, plural, name)
+    .then((res) => {
+      const existingCR = res.body;
+      if (!enableCR) {
+        return false;
+      }
+      return getField(existingCR, appDef.spec.enableCR.field) === appDef.spec.enableCR.value;
+    })
+    .catch(() => false);
+};
+
+const getCSVForApp = (
+  fastify: KubeFastifyInstance,
+  app: OdhApplication,
+): Promise<K8sResourceCommon | undefined> => {
+  if (!app.spec.csvName) {
+    return Promise.resolve(undefined);
+  }
+
+  const subsStatus = getSubscriptions();
+  const subStatus = subsStatus.find((st) => st.installedCSV?.startsWith(app.spec.csvName));
+
+  if (!subStatus) {
+    return Promise.resolve(undefined);
+  }
+
+  const installedCSV = subStatus.installedCSV;
+
+  if (!installedCSV) {
+    return Promise.resolve(undefined);
+  }
+
+  const namespace = subStatus.installPlanRefNamespace;
+  if (!namespace) {
+    return Promise.resolve(undefined);
+  }
+
+  return fastify.kube.customObjectsApi
+    .getNamespacedCustomObject(
+      'operators.coreos.com',
+      'v1alpha1',
+      namespace,
+      'clusterserviceversions',
+      installedCSV,
+    )
+    .then((response) => {
+      const csv = response.body as CSVKind;
+      if (csv.status?.phase === 'Succeeded') {
+        return csv;
+      }
+      return undefined;
+    })
+    .catch((e) => {
+      if (isHttpError(e) && e.statusCode === 404) {
+        // eslint-disable-next-line no-console
+        console.error(e);
+        return undefined;
+      }
+      throw e;
+    });
+};
+
+export const getIsAppEnabled = async (
+  fastify: KubeFastifyInstance,
+  appDef: OdhApplication,
+): Promise<boolean> => {
+  if (appDef.spec.category === 'Red Hat managed') {
+    return true;
+  }
+
+  const enabledCM = await getApplicationEnabledConfigMap(fastify, appDef);
+  if (enabledCM) {
+    return true;
+  }
+  const crEnabled = await getCREnabledForApp(fastify, appDef);
+  if (crEnabled) {
+    return true;
+  }
+
+  if (await getCSVForApp(fastify, appDef)) {
+    return true;
+  }
+
+  return false;
+};
+
+export const getRouteForApplication = async (
+  fastify: KubeFastifyInstance,
+  app: OdhApplication,
+): Promise<string> => {
+  let route = getRouteForClusterId(fastify, app.spec.endpoint);
+  if (route) {
+    return route;
+  }
+
+  route = await getLink(fastify, app.spec.route);
+  if (route) {
+    return route;
+  }
+
+  const operatorCSV = await getCSVForApp(fastify, app);
+  route = await getLink(
+    fastify,
+    app.spec.route,
+    app.spec.routeNamespace || operatorCSV?.metadata.namespace,
+    app.spec.routeSuffix,
+  );
+  if (route) {
+    return route;
+  }
+
+  route = getConsoleLinkRoute(app);
+  if (route) {
+    return route;
+  }
+
+  route = await getServiceLink(fastify, app.spec.serviceName, app.spec.routeSuffix);
+  return route;
 };
 
 const shouldMigrationContinue = async (
