+++ backend/tsconfig.json
@@ -1,17 +1,10 @@
 {
-  "extends": "@odh-dashboard/tsconfig/tsconfig.json",
+  "extends": "@odh-dashboard/tsconfig/tsconfig.node.json",
   "compilerOptions": {
-    "baseUrl": ".",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "paths": {
-      "*": ["node_modules/*"]
-    },
     "noImplicitAny": false,
     "noImplicitReturns": false,
     "strictNullChecks": false,
     "strictFunctionTypes": false,
     "useUnknownInCatchVariables": false
-  },
-  "exclude": ["node_modules", "dist"]
+  }
 }
