--- backend/tsconfig.prod.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "noEmit": false,
-    "emitDeclarationOnly": false,
-    "allowImportingTsExtensions": false,
-    "outDir": "dist"
-  },
-  "include": ["./src/**/*.ts", "./src/**/*.tsx", "./src/**/**.js", "./src/**/*.jsx"],
-  "exclude": [
-    "node_modules",
-    "src/__tests__/"
-  ],
-  "files": []
-}
