+++ package.json
@@ -25,8 +25,8 @@
   ],
   "scripts": {
     "build": "run-p -l build:*",
-    "build:backend": "cd ./backend && npm run build",
-    "build:frontend": "cd ./frontend && npm run build",
+    "build:backend": "npm run build --prefix packages/app-config && npm run build --prefix backend",
+    "build:frontend": "npm run build --prefix frontend",
     "dev": "run-p -l dev:*",
     "dev:backend": "cd ./backend && npm run start:dev",
     "dev:frontend": "cd ./frontend && npm run start:dev",
