--- packages/app-config/src/node/index.ts
@@ -1,2 +1,0 @@
-export * from '../index';
-export * from './module-federation';
