+++ packages/model-registry/upstream/frontend/src/odh/extension-points/deploy.ts
@@ -1,4 +1,3 @@
-import { ProjectKind } from '@odh-dashboard/internal/k8sTypes.js';
 import type { Extension, CodeRef } from '@openshift/dynamic-plugin-sdk';
 import type { ModelDeployPrefillInfo } from '~/odh/hooks/useRegisteredModelDeployPrefillInfo';
